% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{api.get}
\alias{api.get}
\title{API GET call}
\usage{
api.get(endpoint, verbose = TRUE)
}
\arguments{
\item{endpoint}{A character string containing a valid Onboard API endpoint.}

\item{verbose}{Logical. If TRUE, prints response message from API}
}
\value{
A list or data.frame of the API output.
}
\description{
Uses http GET call to return an object from the API.
}
\examples{
\dontrun{ whoami <- api.get('whoami') }

}
