% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{api.status}
\alias{api.status}
\title{Check the status of API connection}
\usage{
api.status()
}
\value{
A character string of the API server status and message.
}
\description{
Provides a status code and message for the API connection.
}
