% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{api_error}
\alias{api_error}
\title{API Error Handler}
\usage{
api_error(status_code)
}
\arguments{
\item{status_code}{An integer, should be from 400-599, in this case the http error code returned by a faulty server request.}
}
\value{
A string with a nicely-formatted error description.
}
\description{
Handles http errors, returning a useful formatted message.
}
