% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{avecClient}
\alias{avecClient}
\title{Exécuter du code dans le contexte d'un client Constellation, et fermer le client par la suite. Nécessite qu'un serveur
Constellation soit déjà activé sur le `port` spécifié.}
\usage{
avecClient(code, port)
}
\arguments{
\item{code}{Le code à exécuter. Ce code doit être une fonction qui prend le `client` Constellation comme unique paramètre.}

\item{port}{Le port du serveur déjà ouvert.}
}
\value{
Le résultat de la fonction `code`.
}
\description{
Exécuter du code dans le contexte d'un client Constellation, et fermer le client par la suite. Nécessite qu'un serveur
Constellation soit déjà activé sur le `port` spécifié.
}
