% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{avecClientEtServeur}
\alias{avecClientEtServeur}
\title{Exécuter du code dans le contexte d'un client Constellation, et fermer le client et le serveur par la suite.}
\usage{
avecClientEtServeur(code, ...)
}
\arguments{
\item{code}{Le code à exécuter. Ce code doit être une fonction qui prend le `client` Constellation comme unique paramètre.}

\item{...}{Paramètres qui seront passés directement à `constellationR::avecServeur`.}
}
\value{
Le résultat de la fonction `code`.
}
\description{
Exécuter du code dans le contexte d'un client Constellation, et fermer le client et le serveur par la suite.
}
