% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serveur.R
\name{avecServeur}
\alias{avecServeur}
\title{Exécuter du code dans le contexte d'un serveur Constellation, et fermer le serveur par la suite.}
\usage{
avecServeur(code, ...)
}
\arguments{
\item{code}{Le code à exécuter. Ce code doit être une fonction qui prend le `port` du serveur comme unique paramètre.}

\item{...}{Arguments qui seront passés directement à `constellationR::lancerServeur`.}
}
\value{
Le résultat de la fonction `code`.
}
\description{
Exécuter du code dans le contexte d'un serveur Constellation, et fermer le serveur par la suite.
}
