% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params.R
\name{building}
\alias{building}
\title{building}
\arguments{
\item{building}{Character or integer corresponding to a single building name or id. Only one value is allowed (functions using this parameter operate on a single building).}
}
\description{
building
}
