% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params.R
\name{building_ids}
\alias{building_ids}
\title{building_ids}
\arguments{
\item{building_ids}{Integer or vector of integers corresponding to building IDs. Unlike `buildings`, these are used directly (e.g. in API endpoint paths) and are not resolved by name.}
}
\description{
building_ids
}
