% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnboardClient-package.R
\name{buildings}
\alias{buildings}
\title{buildings}
\arguments{
\item{buildings}{Integer, character, or a mix of both. Use integers to search by `building_id`, and strings to search by `building name`. Multiple values are allowed.}
}
\description{
buildings
}
