% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_errors}
\alias{check_errors}
\title{Check Errors}
\usage{
check_errors(api_response, verbose = TRUE)
}
\arguments{
\item{api_response}{Response from API Call}

\item{verbose}{Logical. If TRUE, prints response}
}
\description{
Checks Errors from api calls
}
