% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_to_datetime}
\alias{convert_to_datetime}
\title{Convert UNIX timestamp to POSIXct}
\usage{
convert_to_datetime(x)
}
\arguments{
\item{x}{A character or numeric vector of UNIX timestamps.}
}
\value{
A POSIXct vector.
}
\description{
Converts a UNIX timestamp string (with or without milliseconds) to a POSIXct datetime object.
}
