% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging_area.R
\name{create.building}
\alias{create.building}
\title{Create building}
\usage{
create.building(building_name, org_id = NULL, verbose = TRUE)
}
\arguments{
\item{building_name}{Enter the name of the building you want to create on the staging area}

\item{org_id}{(Optional) If missing, the org_id will be grabbed from user's API access}
}
\description{
Creates an empty building on the staging area
}
