% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{create_building}
\alias{create_building}
\title{Create building (Deprecated)}
\usage{
create_building(building_name, org_id = NULL, verbose = TRUE)
}
\arguments{
\item{building_name}{Enter the name of the building you want to create on the staging area}

\item{org_id}{(Optional) If missing, the org_id will be grabbed from user's API access}
}
\description{
Creates an empty building on the staging area
}
