% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{donnéesNuéeÀTrame}
\alias{donnéesNuéeÀTrame}
\title{Transforme des données de nuée de format Constellation en
format de trame de données tibble R.}
\usage{
donnéesNuéeÀTrame(données)
}
\arguments{
\item{données}{Les données provenant de Constellation}
}
\value{
Une trame de données en format tibble R
}
\description{
Transforme des données de nuée de format Constellation en
format de trame de données tibble R.
}
\examples{
données <- jsonlite::fromJSON(
  "{\"données\":[
     {\"col1\":123,
     \"auteur\": \"/orbitdb/zdpuB1wjvzSEsY9YZ4Z2kUEX2DLzwV9G8LCQnzfLccHgY1LdH\"},
     {\"col2\":\"abc\",\"col1\":456,
     \"auteur\": \"/orbitdb/zdpuB1wjvzSEsY9YZ4Z2kUEX2DLzwV9G8LCQnzfLccHgY1LdH\"
     }]}",
  simplifyDataFrame = FALSE
 )
td <- donnéesTableauÀTrame(données["données"])

}
