% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{donnéesTableauÀTrame}
\alias{donnéesTableauÀTrame}
\title{Transforme des données de tableau de format Constellation en
format de trame de données tibble R.}
\usage{
donnéesTableauÀTrame(données)
}
\arguments{
\item{données}{Les données provenant de Constellation}
}
\value{
Une trame de données en format tibble R
}
\description{
Transforme des données de tableau de format Constellation en
format de trame de données tibble R.
}
\examples{
données <- jsonlite::fromJSON("{
  \"données\":[{
    \"col1\":123},{\"col2\":\"abc\",\"col1\":456}
   ]}"
  , simplifyDataFrame = FALSE
 )
td <- donnéesTableauÀTrame(données["données"])

}
