% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params.R
\name{equipment_types}
\alias{equipment_types}
\title{equipment_types}
\arguments{
\item{equipment_types}{Character vector. One or more equipment types (e.g. `"HVAC/AHU"`).}
}
\description{
equipment_types
}
