% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{get_all_units}
\alias{get_all_units}
\title{Units}
\usage{
get_all_units()
}
\value{
A data.frame containing all unit information.
}
\description{
Queries all unit data and information from Onboard's Data Model.
}
