% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_building_info}
\alias{get_building_info}
\title{Building Info}
\usage{
get_building_info(buildings, verbose = TRUE)
}
\arguments{
\item{buildings}{Integer, character, or vectors of those types, providing building id(s) or name(s). You can provide multiple buildings at once.}

\item{verbose}{Logical. If TRUE (default), print status messages.}
}
\value{
A data.frame of building info with two columns, 'id' and 'name'.
}
\description{
Returns building id(s) and name(s) in a named list.
}
