% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{get_deployments}
\alias{get_deployments}
\title{Deployments}
\usage{
get_deployments(org_id)
}
\arguments{
\item{org_id}{organization id}
}
\value{
A data.frame of all deployments.
}
\description{
Get all deployments in your organization.
}
