% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{get_equip_types}
\alias{get_equip_types}
\title{Equipment Types}
\usage{
get_equip_types()
}
\value{
`equip_type`: A dataframe object.
}
\description{
Query all equipment types from Onboard's Data Model.
}
