% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{get_equipment_by_ids}
\alias{get_equipment_by_ids}
\title{Get Equipment by ID}
\usage{
get_equipment_by_ids(equipment_ids, verbose = TRUE)
}
\arguments{
\item{equipment_ids}{Integer or vector of integers. One or more equipment IDs.}

\item{verbose}{Logical; if `TRUE`, prints progress messages (and for `api.request()` specifically, full HTTP request/response details). Defaults to `TRUE`.}
}
\value{
A `data.frame` containing metadata of the requested equipment. Returns an empty list if no matches are found.
}
\description{
Queries equipment metadata by equipment IDs.
}
\examples{
\dontrun{
equipment <- get_equipment_by_ids(c(1000, 1001))

query <- PointSelector()
query$buildings <- 101
query$equipment_types <- 'HVAC/AHU'
selection <- select_points(query)

equipment <- get_equipment_by_ids(selection$equipment)
}

}
