% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{get_orgs}
\alias{get_orgs}
\title{Organizations}
\usage{
get_orgs(id)
}
\arguments{
\item{id}{(Optional) An integer if you want information on a particular entity. Returns all entities unless this argument is provided.}
}
\value{
A data.frame of organization information.
}
\description{
Retrieve Organizations that you have access to.
}
