% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_point_types}
\alias{get_point_types}
\title{Get Point Types}
\usage{
get_point_types()
}
\value{
A data.frame containing all point types.
}
\description{
Queries all point types, measurements and their units from Onboard's Data Model and returns a clean output.
}
