% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{get_published_devices}
\alias{get_published_devices}
\title{Get Published Devices}
\usage{
get_published_devices(building_ids, verbose = TRUE)
}
\arguments{
\item{building_ids}{Character vector or integer corresponding to building ids.}

\item{verbose}{Logical; if `TRUE`, prints verbose output including headers and body content. Defaults to `TRUE`.}
}
\value{
A `data.frame` containing metadata of the published devices for the requested building IDs. Returns an empty list if no matches are found.
}
\description{
Queries published devices by building IDs.
}
\examples{
\dontrun{
devices <- get_published_devices(building_ids = c(427, 428))
}

}
