% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{get_staging_equip}
\alias{get_staging_equip}
\title{Get Staging Equipment
Retrieve all equipment from the staging area for a building.}
\usage{
get_staging_equip(building_id, verbose = TRUE)
}
\arguments{
\item{building_id}{Integer. ID of the building.}

\item{verbose}{Logical; if `TRUE`, prints verbose output including headers and body content. Defaults to `TRUE`.}
}
\value{
A data.frame of staging equipment.
}
\description{
Get Staging Equipment
Retrieve all equipment from the staging area for a building.
}
