% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{get_staging_points}
\alias{get_staging_points}
\title{Get Staging Points
Retrieve all points from the staging area for a building.}
\usage{
get_staging_points(building_id, verbose = TRUE)
}
\arguments{
\item{building_id}{Integer. ID of the building.}

\item{verbose}{Logical; if `TRUE`, prints progress messages (and for `api.request()` specifically, full HTTP request/response details). Defaults to `TRUE`.}
}
\value{
A data.frame of staging points.
}
\description{
Get Staging Points
Retrieve all points from the staging area for a building.
}
