% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Time-Series Data}
\usage{
get_timeseries(
  start_time,
  end_time,
  point_ids,
  units = NULL,
  unit_type = "default"
)
}
\arguments{
\item{start_time}{Start Time in UTC.}

\item{end_time}{End Time in UTC.}

\item{point_ids}{Point IDs for which timeseries data needs to be queried.}

\item{units}{(Optional) A data.frame consisting of preferred units for given measurements}

\item{unit_type}{Provide the unit type: "default" (default) or "raw"}
}
\value{
A wide data.frame of time-series data, with timestamp and all requested point IDs as columns.
}
\description{
Provides clean time-series
}
\examples{
\dontrun{

end_time <- as.POSIXlt(Sys.time(), tz = 'UTC')

start_time <- end_time - lubridate::hours(6)

point_ids <- c(290631, 290643) 

units <- data.frame("temperature" = "k")

timeseries <- get_timeseries(start_time, end_time, point_ids, units)

}

}
