% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serveur.R
\name{installerConstellation}
\alias{installerConstellation}
\title{Installer Constellation sur votre système. Nécessite Node.js (https://nodejs.org/fr) et pnpm (https://pnpm.io/)}
\usage{
installerConstellation()
}
\value{
Rien
}
\description{
Installer Constellation sur votre système. Nécessite Node.js (https://nodejs.org/fr) et pnpm (https://pnpm.io/)
}
