% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serveur.R
\name{obtVersionIPAConstellation}
\alias{obtVersionIPAConstellation}
\title{Obtenir la version installée de l'IPA Constellation}
\usage{
obtVersionIPAConstellation(exe = "constl")
}
\arguments{
\item{exe}{La commande pour lancer Constellation. Uniquement nécessaire pour une installation de Constellation non standard}
}
\value{
La version de l'IPA Constellation
}
\description{
Obtenir la version installée de l'IPA Constellation
}
