% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serveur.R
\name{obtVersionServeur}
\alias{obtVersionServeur}
\title{Obtenir la version installée du serveur Constellation}
\usage{
obtVersionServeur(exe = "constl")
}
\arguments{
\item{exe}{La commande pour lancer Constellation. Uniquement nécessaire pour une installation de Constellation non standard}
}
\value{
La version du serveur Constellation
}
\description{
Obtenir la version installée du serveur Constellation
}
\examples{
# Obtenir la version du serveur installé
v <- obtVersionServeur()
print(v)
}
