% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnboardClient-package.R
\docType{package}
\name{OnboardClient-package}
\alias{OnboardClient}
\alias{OnboardClient-package}
\title{OnboardClient: Bindings for Onboard Data's Building Data API}
\description{
Provides a wrapper for the Onboard Data building data API \url{https://api.onboarddata.io/swagger}. Along with streamlining access to the API, this package simplifies access to sensor time series data, metadata (sensors, equipment, and buildings), and details about the Onboard data model/ontology.
}
\author{
\strong{Maintainer}: Pranay Shah \email{pranay.shah@onboarddata.io}

Other contributors:
\itemize{
  \item Onboard Data \email{support@onboarddata.io} [copyright holder, funder]
}

}
\keyword{internal}
