% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{PointSelector}
\alias{PointSelector}
\title{Point Selector Template}
\usage{
PointSelector()
}
\value{
An empty named list of possible point selection criteria.
}
\description{
Create a query template to select metadata points.
}
\examples{
\dontrun{
query <- PointSelector()

query$buildings <- 101
query$equipment_types <- 'HVAC/AHU'
query$point_types <- c('supply_air_temperature_sensor','supply_air_static_pressure_sensor')
}

}
