% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{préparerNomFoncMessage}
\alias{préparerNomFoncMessage}
\title{Prépare le nom d'une fonction Constellation à être envoyé par WebSocket}
\usage{
préparerNomFoncMessage(nomFonction)
}
\arguments{
\item{nomFonction}{Le nom de la fonction en format vecteur}
}
\value{
Le nom formatté en liste si la longeur > 1, sinon le vecteur original
}
\description{
Prépare le nom d'une fonction Constellation à être envoyé par WebSocket
}
