% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{préparerParamsFoncMessage}
\alias{préparerParamsFoncMessage}
\title{Prépare les paramètres d'une fonction Constellation à être envoyés par WebSocket}
\usage{
préparerParamsFoncMessage(paramètres)
}
\arguments{
\item{paramètres}{La liste des paramètres}
}
\value{
Soit les paramètres eux-mêmes, soit, s'ils sont vides, un `data.frame` vide
}
\description{
Prépare les paramètres d'une fonction Constellation à être envoyés par WebSocket
}
