% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{process_columns}
\alias{process_columns}
\title{Process Columns}
\usage{
process_columns(df)
}
\arguments{
\item{df}{dataframe to parse}
}
\value{
Processed df
}
\description{
Deprecated. Function to process all columns in a dataframe inside a list. Converts list type columns to character. Used in `get_staged_data`
}
