% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{promote}
\alias{promote}
\title{Promote data on Staging Area}
\usage{
promote(building, equip_ids_list = NULL, proceed = NULL, verbose = TRUE)
}
\arguments{
\item{building}{Character vector or integer corresponding to the building name or id.}

\item{equip_ids_list}{A list of all equip_ids to promote from staging to the live building}

\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}
}
\value{
(Conditional) A named list with result output of promotion.
}
\description{
Promote valid data on the staging area to the live building.
}
