% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging_area.R
\name{promote_staged_data}
\alias{promote_staged_data}
\title{Promote data on Staging Area}
\usage{
promote_staged_data(building, data_to_promote)
}
\arguments{
\item{building}{Enter building id or name. Note: If you enter multiple building ids or names, only the first entry is considered}

\item{data_to_promote}{(Optional) If missing, all valid topics are promoted. A dataframe containing e.equip_id & p.topic columns}
}
\description{
Promote valid data on the staging area to the live building
}
