% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{résoudreTypeFonction}
\alias{résoudreTypeFonction}
\title{Essaie de deviner de quel type est une fonction Constellation.}
\usage{
résoudreTypeFonction(nomFonction)
}
\arguments{
\item{nomFonction}{Le nom de la fonction selon la documentation Constellation (https://docu.réseau-constellation.ca), par exemple, "obtIdCompte" ou "variables.créerVariable"}
}
\value{
Le type de fonction, soit "suivi", "action", ou "recherche".
}
\description{
Essaie de deviner de quel type est une fonction Constellation.
}
\examples{
typeFonction <- résoudreTypeFonction("recherche.rechercherVariablesSelonNom")
typeFonction == "recherche"

}
