% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{search_by_org}
\alias{search_by_org}
\title{Search by Org}
\usage{
search_by_org(data, org = NULL)
}
\arguments{
\item{data}{A data frame containing an `org_id` column. If `org_name` is missing, it will be added via API.}

\item{org}{Organization ID (numeric) or name (character)}
}
\value{
Filtered data frame with `org_name` included.
}
\description{
Search by Org
}
