% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{search_orgs}
\alias{search_orgs}
\title{Search Organizations}
\usage{
search_orgs(orgs = NULL, vebrose = TRUE)
}
\arguments{
\item{orgs}{Integer, character, or a mix of both. Use integers to search by `org_id`, and strings to search by `org_name`. Multiple values are allowed.}
}
\value{
A dataframe containing the matched org(s).
}
\description{
Search Organizations
}
