% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{unpromote}
\alias{unpromote}
\title{Unpromote Data from the live Building}
\usage{
unpromote(
  building,
  data_to_unpromote,
  unpromote_type = "points",
  proceed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{building}{Character vector or integer corresponding to the building name or id. If you enter multiple building ids or names, only the first entry will be considered.}

\item{unpromote_type}{Type of data to unpromote. "equipment", "points"(default) or "relationships_only"}

\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}

\item{verbose}{Logical. If TRUE (default), prints status and progress messages.}

\item{data_to_unpromotw}{A data.frame containing columns 'e.equipment_id' and/or 'p.point_id'.}
}
\value{
(Conditional) A named list containing errors that may have occurred during data promotion.
}
\description{
Unpromote Data from the live Building
}
