% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging_area.R
\name{unpromote_data}
\alias{unpromote_data}
\title{Unpromote Data from the live Building}
\usage{
unpromote_data(building, data_to_unpromote, verbose = TRUE)
}
\arguments{
\item{building}{Character vector or integer corresponding to the building name or id. If you enter multiple building ids or names, only the first entry is considered.}

\item{data_to_unpromote}{A data.frame containing columns 'e.equip_id' & 'p.topic'.}

\item{verbose}{Logical. If TRUE (default), prints status and progress messages.}
}
\value{
(Conditional) A named list containing errors that may have occurred during data promotion.
}
\description{
Unpromote Data from the live Building
}
