% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{update_staging_equipment}
\alias{update_staging_equipment}
\title{Update Staging Equipment}
\usage{
update_staging_equipment(
  building,
  staging_update,
  proceed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{building}{Character vector or integer corresponding to the building name or id.}

\item{staging_update}{A data.frame to upload to the staging area. Must contain name (equip_name) column. equipment_type_tag_name & new_name are optional columns}

\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}
}
\value{
Result output of the update
}
\description{
Update equipment on the staging area.
}
