% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnboardClient-package.R
\name{verbose}
\alias{verbose}
\title{verbose}
\arguments{
\item{verbose}{Logical; if `TRUE`, prints verbose output including headers and body content. Defaults to `TRUE`.}
}
\description{
verbose
}
