% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{api.access}
\alias{api.access}
\title{Access API keys and URL from System Environment}
\usage{
api.access()
}
\value{
A named list of api data, including 'api_url' and 'api_key'.
}
\description{
Returns the api_url and api_keys.
}
