% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{api.delete}
\alias{api.delete}
\title{API DELETE call}
\usage{
api.delete(endpoint, json_body = NULL, verbose = TRUE)
}
\arguments{
\item{endpoint}{A character string containing a valid Onboard API endpoint.}

\item{json_body}{A JSON payload to give to the POST call.}
}
\description{
Uses http DELETE call to post objects to the API.
}
