% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{api.get}
\alias{api.get}
\title{API GET call}
\usage{
api.get(endpoint)
}
\arguments{
\item{endpoint}{A character string containing a valid Onboard API endpoint.}
}
\value{
An R object of class `list` or `data.frame`
}
\description{
Uses http GET call to return an object from the API.
}
\examples{
\dontrun{ whoami <- api.get('whoami') }

}
