% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{api.post}
\alias{api.post}
\title{API POST call}
\usage{
api.post(endpoint, json_body, output = "list")
}
\arguments{
\item{endpoint}{A character string containing a valid Onboard API endpoint.}

\item{json_body}{A JSON payload to give to the POST call.}

\item{output}{if "list" (default), it returns the api output as a list object. If "dataframe", it returns the api output as a dataframe object}
}
\value{
An R object of `list` or `data.frame` class
}
\description{
Uses http POST call to post objects to the API
}
