% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{api.post}
\alias{api.post}
\title{API POST call}
\usage{
api.post(endpoint, json_body, upload_path = NULL, output = "list")
}
\arguments{
\item{endpoint}{A character string containing a valid Onboard API endpoint.}

\item{json_body}{A JSON payload to give to the POST call.}

\item{upload_path}{(Optional) A character string containing the file path for the file to upload}

\item{output}{A character string, either "list" (default) or "dataframe", to specify the API output format.}
}
\value{
A list or data.frame of the API output.
}
\description{
Uses http POST call to post objects to the API.
}
