% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{api.setup}
\alias{api.setup}
\title{Set up Onboard API keys and URL in system environment}
\usage{
api.setup(api_env = "prod", key = NULL, token = NULL, verbose = TRUE)
}
\arguments{
\item{api_env}{Provide the API client name. Default is `prod`}

\item{key}{Option to include your API keys in the argument. It is NULL by default}

\item{token}{Option to include your Auth token in the argument. It is NULL by default}

\item{verbose}{Logical. If TRUE, prints the API status.}
}
\value{
No return value, sets API URL and API key in the system environment.
}
\description{
Set the Onboard API URL and API keys in the system environment.
}
