% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{api.setup}
\alias{api.setup}
\title{Set up Onboard API keys and URL in system environment}
\usage{
api.setup(api_type)
}
\arguments{
\item{api_type}{Optional. If `dev`, will setup api url and keys for acessing Onboard's DEV API.}
}
\description{
Sets the onboard api url and api keys in the system environment.
}
