% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{Client}
\alias{Client}
\title{Représent, en R, un client Constellation.}
\description{
Cette classe se connecte à un serveur Constellation déjà ouvert. Nous vous recommandons de ne pas l'utiliser directement,
mais plutôt d'appeler `constellationR::avecClientEtServeur`, ou bien `constellationR::avecClient`, lesquels
s'occuperont de la création et de la fermeture du client pour vous.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{port}}{Le numéro du port local sur lequel le serveur est ouvert, et auquel le client se connectera.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{port}}{Le numéro du port local sur lequel le serveur est ouvert, et auquel le client se connectera.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClientConstellation-new}{\code{Client$new()}}
\item \href{#method-ClientConstellation-action}{\code{Client$action()}}
\item \href{#method-ClientConstellation-suivre}{\code{Client$suivre()}}
\item \href{#method-ClientConstellation-rechercher}{\code{Client$rechercher()}}
\item \href{#method-ClientConstellation-appeler}{\code{Client$appeler()}}
\item \href{#method-ClientConstellation-enregistrerÉcoute}{\code{Client$enregistrerÉcoute()}}
\item \href{#method-ClientConstellation-fermer}{\code{Client$fermer()}}
\item \href{#method-ClientConstellation-clone}{\code{Client$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConstellation-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConstellation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$new(port)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConstellation-action"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConstellation-action}{}}}
\subsection{Method \code{action()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$action(fonction, paramètres = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConstellation-suivre"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConstellation-suivre}{}}}
\subsection{Method \code{suivre()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$suivre(fonction, paramètres = NULL, nomArgFonction = "f")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConstellation-rechercher"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConstellation-rechercher}{}}}
\subsection{Method \code{rechercher()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$rechercher(fonction, paramètres, nomArgFonction = "f")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConstellation-appeler"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConstellation-appeler}{}}}
\subsection{Method \code{appeler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$appeler(fonction, paramètres = NULL, nomArgFonction = "f")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConstellation-enregistrerÉcoute"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConstellation-enregistrerÉcoute}{}}}
\subsection{Method \code{enregistrerÉcoute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$enregistrerÉcoute(id, résoudre, rejeter, f = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConstellation-fermer"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConstellation-fermer}{}}}
\subsection{Method \code{fermer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$fermer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConstellation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConstellation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
