% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{demote}
\alias{demote}
\title{Demote Data from the live Building}
\usage{
demote(
  building,
  equipment_ids = NULL,
  point_ids = NULL,
  proceed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{building}{Character vector or integer corresponding to the building name or id. If you enter multiple building ids or names, only the first entry will be considered.}

\item{equipment_ids}{NULL (default) or A vector of equipment_ids to demote. Provide atleast one of equipment_ids or point_ids}

\item{point_ids}{NULL(default) A vector of equipment_ids to demoted. Provide atleast one of equipment_ids or point_ids}

\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}

\item{verbose}{Logical. If TRUE (default), prints status and progress messages.}
}
\value{
(Conditional) A named list containing errors that may have occurred during data promotion.
}
\description{
Demote Data from the live Building
}
