% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_building_info}
\alias{get_building_info}
\title{Building Info}
\usage{
get_building_info(buildings)
}
\arguments{
\item{buildings}{A character vector or integer. Provide either building id or name. You can provide multiple buildings at one}
}
\value{
id: (Integer) Building ID; name: (Character) Building Name
}
\description{
Retrieves building id and name
}
\examples{
\dontrun{
get_building_info(buildings=c(427,"Laboratory"))
}

}
