% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_deployments}
\alias{get_deployments}
\title{Deployments}
\usage{
get_deployments(org = NULL)
}
\arguments{
\item{org}{Enter org_id (integer) or org_name (text)}
}
\value{
A data.frame of all deployments.
}
\description{
Get all deployments in your organization.
}
