% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{get_metadata}
\alias{get_metadata}
\title{GET Metadata}
\usage{
get_metadata(buildings = NULL, query = NULL, verbose = TRUE)
}
\arguments{
\item{buildings}{Integer, character, or a mix of both. Use integers to search by `building_id`, and strings to search by `building name`. Multiple values are allowed.}

\item{query}{Building parameters query from the point selector.}

\item{verbose}{Logical; if `TRUE`, prints verbose output including headers and body content. Defaults to `TRUE`.}
}
\value{
A data.frame of clean metadata for the requested points.
}
\description{
Retrieves live points and equipment for a given building or selection and outputs a clean metadata data.frame.
}
\examples{
\dontrun{
metadata <- get_metadata(buildings = c(427, "Laboratory"))

OR

query <- PointSelector()
query$buildings <- 427
query$equipment_types <- 'HVAC/AHU'
query$point_types <- c('supply_air_temperature_sensor','supply_air_static_pressure_sensor')

metadata <- get_metadata(query = query)
}
}
