% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Metadata}
\usage{
get_metadata(buildings, selection)
}
\arguments{
\item{buildings}{A character vector or integer. Provide either building id or name. You can provide multiple buildings at one}

\item{selection}{Selection list from point selector}
}
\description{
Retrieves points and equipment for a given building or selection and outputs a clean metadata dataframe
}
\examples{
\dontrun{
metadata <- get_metadata(buildings=c(427,"Laboratory"))

OR

query <- PointSelector()

query$buildings <- 427
query$equipment_types <- 'ahu'
query$point_types <- c('Supply Air Temperature','Supply Air Static Pressure')

selection <- select_points(query)

metadata <- get_metadata(selection)
}

}
