% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_orgs}
\alias{get_orgs}
\title{Retrieve Organizations that you have access to.}
\usage{
get_orgs(id)
}
\arguments{
\item{id}{(Optional) An integer if you want information on a particular entity. Returns all entities unless this argument is provided.}
}
\description{
Retrieve Organizations that you have access to.
}
