% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{get_points_by_ids}
\alias{get_points_by_ids}
\title{Get Points by ID}
\usage{
get_points_by_ids(point_ids, verbose = TRUE)
}
\arguments{
\item{point_ids}{Integer or vector of integers. One or more point IDs.}

\item{verbose}{Logical; if `TRUE`, prints progress messages (and for `api.request()` specifically, full HTTP request/response details). Defaults to `TRUE`.}
}
\value{
A `data.frame` containing metadata of the requested points. Returns an empty `data.frame` if no points are found.
}
\description{
Queries point metadata by a list of point IDs.
}
\examples{
\dontrun{
points <- get_points_by_ids(c(10000, 10001))

query <- PointSelector()
query$buildings <- 101
query$equipment_types <- 'HVAC/AHU'
query$point_types <- c('supply_air_temperature_sensor', 'supply_air_static_pressure_sensor')

selection <- select_points(query)
points <- get_points_by_ids(selection$points)
}

}
