% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging_area.R
\name{get_staged_data}
\alias{get_staged_data}
\title{Get Staged Data}
\usage{
get_staged_data(building)
}
\arguments{
\item{building}{Enter building id or name. Note: If you enter multiple building ids or names, only the first entry is considered}
}
\description{
Gets metadata from the staging area
}
