% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{get_staging_data}
\alias{get_staging_data}
\title{Get Staging Data
Retrieve metadata (points, equipment, devices) from the staging area for one or more buildings.}
\usage{
get_staging_data(buildings, verbose = TRUE)
}
\arguments{
\item{buildings}{Integer, character, or a mix of both. Use integers to search by `building_id`, and strings to search by `building name`. Multiple values are allowed.}

\item{verbose}{Logical; if `TRUE`, prints progress messages (and for `api.request()` specifically, full HTTP request/response details). Defaults to `TRUE`.}
}
\value{
A data.frame containing combined metadata from the staging area.
}
\description{
Get Staging Data
Retrieve metadata (points, equipment, devices) from the staging area for one or more buildings.
}
