% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{get_staging_data}
\alias{get_staging_data}
\title{Get Staging Data}
\usage{
get_staging_data(buildings, verbose = TRUE)
}
\arguments{
\item{buildings}{Enter building names or ids. You can enter multiple building ids or names as vectors}

\item{verbose}{Logical. If TRUE (default), prints status and progress messages.}
}
\value{
A data.frame of metadata from the staging area.
}
\description{
Gets metadata from the staging area.
}
