% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{get_staging_devices}
\alias{get_staging_devices}
\title{Get Staging Devices}
\usage{
get_staging_devices(building_id, verbose = TRUE)
}
\arguments{
\item{building_id}{Integer corresponding to the building id}
}
\value{
A data.frame of staging devices
}
\description{
Get unfiltered staging details on the building devices
}
