% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{get_staging_equip}
\alias{get_staging_equip}
\title{Get Staging Equip}
\usage{
get_staging_equip(building_id, verbose = TRUE)
}
\arguments{
\item{building_id}{Integer corresponding to the building id}
}
\value{
A data.frame of staging equipment
}
\description{
Get unfiltered staging details on the building equipment
}
