% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{get_staging_points}
\alias{get_staging_points}
\title{Get Staging Points}
\usage{
get_staging_points(building_id, verbose = TRUE)
}
\arguments{
\item{building_id}{Integer corresponding to the building id}
}
\value{
A data.frame of staging points
}
\description{
Get unfiltered staging details on the building points
}
