% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Time-Series Data
Returns cleaned, wide-format time-series data with one column per point ID.}
\usage{
get_timeseries(
  start_time,
  end_time,
  point_ids,
  units = NULL,
  aggregations = NULL,
  resample_mins = NULL,
  verbose = TRUE
)
}
\arguments{
\item{start_time}{Start time in UTC.}

\item{end_time}{End time in UTC.}

\item{point_ids}{Numeric. Vector of point IDs to query.}

\item{units}{Optional list of preferred units for measurements.}

\item{aggregations}{Optional list of preferred aggregations. Accepted values: max, min, mean(default), sum}

\item{resample_mins}{Optional numeric value of minutes for re-sampling data}

\item{verbose}{Logical; if `TRUE`, prints progress messages (and for `api.request()` specifically, full HTTP request/response details). Defaults to `TRUE`.}
}
\value{
A wide data.frame of time-series data, with timestamp and all requested point IDs as columns.
}
\description{
Time-Series Data
Returns cleaned, wide-format time-series data with one column per point ID.
}
\examples{
\dontrun{

end_time <- as.POSIXlt(Sys.time(), tz = 'UTC')

start_time <- end_time - lubridate::hours(6)

point_ids <- c(290631, 290643) 

units <- list("temperature" = "k")

aggregations <- list("energy"="max")

resample_mins = 120

timeseries <- get_timeseries(start_time, end_time, point_ids, units)

}

}
