% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Time-Series Data
Returns cleaned, wide-format time-series data with one column per point ID.}
\usage{
get_timeseries(
  start_time,
  end_time,
  point_ids,
  units = NULL,
  unit_type = "default",
  verbose = TRUE
)
}
\arguments{
\item{start_time}{Start time in UTC.}

\item{end_time}{End time in UTC.}

\item{point_ids}{Numeric. Vector of point IDs to query.}

\item{units}{Optional data.frame. Preferred units for measurements.}

\item{unit_type}{Character. Use "default" (default) or "raw" values.}

\item{verbose}{Logical; if `TRUE`, prints verbose output including headers and body content. Defaults to `TRUE`.}
}
\value{
A wide data.frame of time-series data, with timestamp and all requested point IDs as columns.
}
\description{
Time-Series Data
Returns cleaned, wide-format time-series data with one column per point ID.
}
\examples{
\dontrun{

end_time <- as.POSIXlt(Sys.time(), tz = 'UTC')

start_time <- end_time - lubridate::hours(6)

point_ids <- c(290631, 290643) 

units <- data.frame("temperature" = "k")

timeseries <- get_timeseries(start_time, end_time, point_ids, units)

}

}
