% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_users}
\alias{get_users}
\title{Users}
\usage{
get_users(org = NULL)
}
\arguments{
\item{org}{Organization ID (numeric) or name (character)}
}
\value{
A data.frame of all user information.
}
\description{
Retrieve all user info in your organization.
}
