% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_users}
\alias{get_users}
\title{Users}
\usage{
get_users(orgs = NULL, verbose = TRUE)
}
\arguments{
\item{orgs}{(Optional) Integer, character, or a mix of both. Use integers to search by `org_id`, and strings to search by `org_name`. Multiple values are allowed.}

\item{verbose}{Logical; if `TRUE`, prints verbose output including headers and body content. Defaults to `TRUE`.}
}
\value{
A data.frame of all user information.
}
\description{
Retrieve all user info in your organization.
}
