% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_users}
\alias{get_users}
\title{Users}
\usage{
get_users(id)
}
\arguments{
\item{id}{(Optional) An integer if you want information on a particular entity. Will return all entities if argument is empty}
}
\description{
Retrieve all user info in your organization.
}
