% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serveur.R
\name{lancerServeur}
\alias{lancerServeur}
\title{Lancer un serveur Constellation local}
\usage{
lancerServeur(port = NULL, dossier = NULL, exe = "constl")
}
\arguments{
\item{port}{Le numéro du port sur lequel le port sera connecté}

\item{dossier}{Le dossier du compte Constellation à utiliser (optionnel)}

\item{exe}{La commande pour lancer Constellation. Uniquement nécessaire pour une installation de Constellation non standard}
}
\value{
Le numéro de port sur lequel le le serveur écoute désormais, et une fonction à appeler pour fermer le serveur
}
\description{
Lancer un serveur Constellation local
}
