% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lancerServeur}
\alias{lancerServeur}
\title{Lancer un serveur Constellation local}
\usage{
lancerServeur(port = NULL, autoinstaller = TRUE, exe = "constl")
}
\arguments{
\item{port}{Le numéro du port sur lequel le port sera connecté}

\item{autoinstaller}{Si on devrait installer Constellation pour vous, si ce n'est pas déjà fait}

\item{exe}{La commande pour lancer Constellation. Uniquement nécessaire pour une installation de Constellation non standard}
}
\value{
Le numéro de port sur lequel le le serveur écoute désormais
}
\description{
Lancer un serveur Constellation local
}
\examples{
# Lancer sans port défini
port <- lancerServeur()
# Lancer sur un port prédéfini (il doit être libre)
lancerServeur(port=5001)
}
