% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{nested_list_to_df}
\alias{nested_list_to_df}
\title{Nested List to DF}
\usage{
nested_list_to_df(nested_list)
}
\arguments{
\item{nested_list}{time-series output return from /timeseries API endpoint}
}
\value{
Clean dataframe
}
\description{
Vectorized approach to convert a deeply nested into a dataframe. This function is customized only to work with time-series output. Used in `get_timeseries_raw`
}
