% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnboardClient-package.R
\docType{package}
\name{OnboardClient-package}
\alias{OnboardClient}
\alias{OnboardClient-package}
\title{OnboardClient: OnboardClient}
\description{
Provides R bindings to Onboard Data's building data API.
}
\author{
\strong{Maintainer}: Pranay Shah \email{pranay.shah@onboarddata.io}

Authors:
\itemize{
  \item Christopher Dudas-Thomas \email{christopher@onboarddata.io}
}

Other contributors:
\itemize{
  \item Onboard Data \email{support@onboarddata.io} [copyright holder, funder]
}

}
\keyword{internal}
