% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params.R
\name{orgs}
\alias{orgs}
\title{orgs}
\arguments{
\item{orgs}{(Optional) Integer, character, or a mix of both. Use integers to search by `org_id`, and strings to search by org name, short name, or abbreviation. Multiple values are allowed.}
}
\description{
orgs
}
