% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnboardClient-package.R
\name{orgs}
\alias{orgs}
\title{orgs}
\arguments{
\item{orgs}{Integer, character, or a mix of both. Use integers to search by `org_id`, and strings to search by `org_name`. Multiple values are allowed.}
}
\description{
orgs
}
