% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params.R
\name{proceed}
\alias{proceed}
\title{proceed}
\arguments{
\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}
}
\description{
proceed
}
