% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{promote}
\alias{promote}
\title{Promote data on Staging Area}
\usage{
promote(
  building,
  staging_data_to_promote = NULL,
  proceed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{building}{Character vector or integer corresponding to the building name or id. If you enter multiple building ids or names, only the first entry will be considered.}

\item{staging_data_to_promote}{(Optional) If missing, all valid topics are promoted. A data.frame containing columns 'e.equip_id' & 'p.topic'.}

\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}

\item{verbose}{Logical. If TRUE (default), prints status and progress messages.}
}
\value{
(Conditional) A named list containing errors that may have occurred during data promotion.
}
\description{
Promote valid data on the staging area to the live building.
}
