% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{publish}
\alias{publish}
\title{Publish data on Staging Area}
\usage{
publish(
  building,
  equipment = NULL,
  topics = NULL,
  proceed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{building}{Character vector or integer corresponding to the building name or id.}

\item{equipment}{A vector containing all equip_ids to Publish from staging to the live building}

\item{topics}{(Optional) A vector containing all topics to publish for the corresponding equip_ids}

\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}
}
\value{
(Conditional) A named list with result output of promotion.
}
\description{
Publish valid data on the staging area to the live building.
}
