% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{search_buildings}
\alias{search_buildings}
\title{Search Buildings}
\usage{
search_buildings(buildings = NULL, verbose = TRUE)
}
\arguments{
\item{buildings}{Integer, character, or vectors of those types, providing building id(s) or name(s). You can provide multiple buildings at once. Only provide integers if you want to search by building_id}

\item{verbose}{Logical. If TRUE (default), print status messages.}
}
\value{
A dataframe with the building info results
}
\description{
Search any buildings in your org
}
\examples{
\dontrun{

Integer values will search by building ids only
buildings = c(427, 428)

Combination of integer and characters will search ids and names of buildings
buildings=c(427,"Laboratory")

search_building(buildings = buildings)
}

}
