% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{search_buildings}
\alias{search_buildings}
\title{Search Buildings in Your Organization}
\usage{
search_buildings(orgs = NULL, buildings = NULL, verbose = TRUE)
}
\arguments{
\item{orgs}{(Optional) Integer, character, or a mix of both. Use integers to search by `org_id`, and strings to search by `org_name`. Multiple values are allowed.}

\item{buildings}{Integer, character, or a mix of both. Use integers to search by `building_id`, and strings to search by `building name`. Multiple values are allowed.}

\item{verbose}{Logical; if `TRUE`, prints verbose output including headers and body content. Defaults to `TRUE`.}
}
\value{
A dataframe containing the matched building(s).
}
\description{
Search buildings in your organization by ID or name.
}
\examples{
\dontrun{
# Search by building IDs
search_buildings(buildings = c(427, 428))

# Search by name
search_buildings(buildings = "Laboratory")

# Mixed search
search_buildings(buildings = c(427, "Laboratory"))
}

}
