% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{search_buildings}
\alias{search_buildings}
\title{Search Buildings}
\usage{
search_buildings(buildings = NULL, verbose = TRUE)
}
\arguments{
\item{buildings}{Integer, character, or vectors of those types, providing building id(s) or name(s). You can provide multiple buildings at once.}

\item{verbose}{Logical. If TRUE (default), print status messages.}
}
\value{
A dataframe with the building info results
}
\description{
Search any buildings in your org
}
