% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{search_by_org}
\alias{search_by_org}
\title{Search by Org}
\usage{
search_by_org(data, org = NULL)
}
\arguments{
\item{data}{A dataframe to filter}

\item{org}{An id (integer) or name (text) of the organization}
}
\description{
Search by Org
}
