% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{search_orgs}
\alias{search_orgs}
\title{Search Organizations}
\usage{
search_orgs(orgs = NULL, verbose = TRUE)
}
\arguments{
\item{orgs}{(Optional) Integer, character, or a mix of both. Use integers to search by `org_id`, and strings to search by org name, short name, or abbreviation. Multiple values are allowed.}

\item{verbose}{Logical; if `TRUE`, prints progress messages (and for `api.request()` specifically, full HTTP request/response details). Defaults to `TRUE`.}
}
\value{
A dataframe containing the matched org(s).
}
\description{
Search Organizations
}
