% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{select_points}
\alias{select_points}
\title{Select Points Based on Query}
\usage{
select_points(query, verbose = TRUE)
}
\arguments{
\item{query}{A named list, typically created using `PointSelector()`, defining filter criteria such as buildings, point types, or equipment.}

\item{verbose}{Logical; if `TRUE`, prints verbose output including headers and body content. Defaults to `TRUE`.}
}
\value{
A named list of all the selected buildings, equipment IDs, equipment_types, point_types & point IDs for that category.
}
\description{
Use a query generated from `PointSelector()` to fetch matching point metadata via a POST API request.
}
\examples{
\dontrun{
query <- PointSelector()
query$buildings <- 427
query$equipment_types <- 'HVAC/AHU'
query$point_types <- c('supply_air_temperature_sensor','supply_air_static_pressure_sensor')

points <- select_points(query)
}

}
