% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{unpromote_data}
\alias{unpromote_data}
\title{Unpromote Data from the live Building}
\usage{
unpromote_data(
  building,
  points_to_unpromote = NULL,
  equipment_to_unpromote = NULL,
  proceed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{building}{Character vector or integer corresponding to the building name or id. If you enter multiple building ids or names, only the first entry will be considered.}

\item{points_to_unpromote}{A data.frame containing columns 'e.equipment_id' & 'p.point_id'.}

\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}

\item{verbose}{Logical. If TRUE (default), prints status and progress messages.}
}
\value{
(Conditional) A named list containing errors that may have occurred during data promotion.
}
\description{
Unpromote Data from the live Building
}
