% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{unpublish}
\alias{unpublish}
\title{Unpublish Data from the live Building}
\usage{
unpublish(
  building,
  equipment_ids = NULL,
  point_ids = NULL,
  point_equipment_relationships = NULL,
  proceed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{building}{Character vector or integer corresponding to a building name or id.}

\item{equipment_ids}{NULL (default) or A vector of equipment_ids to unpublish. Provide atleast one of equipment_ids or point_ids}

\item{point_ids}{NULL(default) A vector of equipment_ids to unpublishd. Provide atleast one of equipment_ids or point_ids}

\item{point_equipment_relationships}{NULL(default) A data.frame containing equipment_id and point_id relationships to unpublish}

\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}
}
\value{
(Conditional) A named list containing errors that may have occurred during data promotion.
}
\description{
Unpublish Data from the live Building
}
