% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{update_staging_equip}
\alias{update_staging_equip}
\title{Update Staging Equip}
\usage{
update_staging_equip(building, staging_equip, proceed = NULL, verbose = TRUE)
}
\arguments{
\item{building}{Character or integer corresponding to a single building name or id. Only one value is allowed (functions using this parameter operate on a single building).}

\item{staging_equip}{A data.frame to upload to the staging area. Must contain name (equip_name) column. equipment_type_tag_name, equipment_type_confidence & new_name are optional columns}

\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}
}
\value{
Result output of the update
}
\description{
Update equipment on the staging area.
}
