% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{update_staging_points}
\alias{update_staging_points}
\title{Update Staging Points}
\usage{
update_staging_points(building, staging_points, proceed = NULL, verbose = TRUE)
}
\arguments{
\item{building}{Character vector or integer corresponding to the building name or id.}

\item{staging_points}{A data.frame to upload to the staging area. Must contain equip_names and topic columns. point_type_tag_name, point_type_confidence and raw_unit are optional columns}

\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}
}
\value{
Result output of the update
}
\description{
Update points on the staging area.
}
