% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{upload_staging}
\alias{upload_staging}
\title{Upload to Staging Area}
\usage{
upload_staging(building, staging_data, proceed = NULL, verbose = TRUE)
}
\arguments{
\item{building}{Character vector or integer corresponding to the building name or id. If you enter multiple building ids or names, only the first entry will be considered.}

\item{staging_data}{A data.frame to upload to the staging area. Must contain e.equip_id and p.topic columns.}

\item{proceed}{(Optional) Logical argument indicating whether to proceed operation without asking for explicit user input. Useful for scripting}

\item{verbose}{Logical. If TRUE (default), prints status and progress messages.}
}
\value{
A named list containing any errors that may have occured during data upload.
}
\description{
Uploads data to the staging area.
}
