% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging_area.R
\name{upload_staging}
\alias{upload_staging}
\title{Upload to Staging Area}
\usage{
upload_staging(building, data_to_upload, skip_topics)
}
\arguments{
\item{data_to_upload:}{A dataframe to upload to the staging area. Must contain e.equip_id and p.topic columns}

\item{skip_topics:}{Logical. If True, the uploaded topics will be assigned `__SKIP__` equip_id}
}
\description{
Uploads data to the staging area
}
