% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-params.R
\name{verbose}
\alias{verbose}
\title{verbose}
\arguments{
\item{verbose}{Logical; if `TRUE`, prints progress messages (and for `api.request()` specifically, full HTTP request/response details). Defaults to `TRUE`.}
}
\description{
verbose
}
